Added support for eventlog (#129)
Eventlog is append-only log. 
Will be used in
- Olympus as a log of backed modules
- Proxy as a log of backed modules
- Proxy as a log of registered cache misses (clearable after read)

Multiple backing storages should be supported, for now, mongo is implemented. 
Preparation for Olympus reader, which is remote reader of Olympus event log, is in place without implementation, this implementation is for proxies to read logs from Olympus, and Olympus instances between each other (Pull synchronization)