Several minor UI improvements (#57)
* style: display invalid dates as (unknown date)

Before, they were shown as "12/31/1969, 7:00:00 PM". The new text is
shown in dark gray. I've also re-formatted style.cljs.

* style: alternate gray and white on pages table

* feat: no blank screen during JavaScript loading

Shows a "Loading script..." while the JS loads, as that can take a bit.
Otherwise, the user would just see a blank screen. It also informs the
user if the JS is loading currently or the database is loading
currently.

* style: add a small gap between left sidebar links

* refactor: order requires alphabetically

My linting tool was complaining about unordered requires,
this orders them.

* fix: remove Roam reference while loading

* refactor: use merging functions instead of classes

Co-authored-by: Jeroen van Dijk <jeroentjevandijk@gmail.com>