fix(parsing): fix red parse failures in block contents and refactor parser
The parse failures were because lines beginning with punctuation such as `{` were neither word characters (`\w`) or space characters (`\s`). I fixed that with the `any-char` rule.

The `any-char` rule matches `\w\W` instead of `.` to ensure it also matches newlines, no matter how Clojure and ClojureScript’s regexes work. This is desired in normal text, but probably not in links and hash-tags. That’s one reason for my comment about planning to move away from the `any-chars` rule.

I could have fixed it by just changing the `c` regexp to `#'(\\w|\\W)+'`, but that would lead to other problems down the line. See https://github.com/Engelberg/instaparse#regular-expressions-a-word-of-warning. The downside of doing it repetition in the parser is that I need to combine the individually-matched characters in the transformer later, with `join` and `combine-adjacent-strings`.

Other changes:

- Use clearer names for the parts of the parser
- Use `defparser` to precompile the parser for ClojureScript (following the advice in https://github.com/Engelberg/instaparse#defparser)
- Add classes to output elements to aid in future styling and debugging
- Move the block-ref parsing rule closer to the block-link parsing rule
- Link to the Instaparse docs to make contributing to this file easier