refactor(parsing): make the parser testable and add a few tests
The tests run with Clojure, not ClojureScript, so the code to test had to be moved to a `.cljc` file to be available in both environments.

Only the parsing code, not the transforming/rendering code, was moved into the new file. That’s because the rendering code uses the re-frame and reitit libraries to return its value, so it can’t be tested without mocking them somehow. I’m not sure if testing rendered HTML is even desirable (though testing other aspects of the transformation might be nice).