feat(parsing): parse formatting syntax within url-link-text
Allow writing bold-formatted text in a link with `**foo**`, or writing a literal `]` with `\]`.

As the parsing of more formatting syntax such as `~~strikethrough~~` is implemented, this parse rule might need to be updated to accept that syntax as well. Only some of the new syntax will be allowed in link text.

The backslash-escaping was taken from Markdown. Roam doesn’t support that syntax (yet?), but I think Athens should.