Check for stale parent when deleting block
Holding down backspace when deleting text can enqueue multiple
:backspace reg-event-fxes. The first will correctly delete and merge
the parent blocks, but the next events in the queue will still be processed.
This is a surgical fix that just checks for an invalid parent and
fixes #567, but a more heavy-handed approach may be appropriate to
handle other unreported edge-cases of event queueing.