fix(router): Refactor navigation page-title to avoid nil error
Original code alows executing a pull for `[:block/uid nil]`
which is not a valid lookup ref (and throws an exception
in DataScript). It was not a problem up to now, because
Posh does allow this unspecified behavior (and just returns nil).

Refactored the code to separate page-title and navigation concerns.