perf(db): Replace posh with reagent.ratom (kudos @lambduhh)
Solution was originally proposed by @lambduhh in
https://github.com/athensresearch/athens/pull/665.
See PR discussion for details why posh is considered
slow and a mismatch for Athens UI performance right now.

This version of the solution introduces a new namespace
`athens.posh` as a form of indirection. This way the rest
of the codebase does not need any further changes (aside from
the namespace require) and also allows easy switching between
the two different databases to compare performance and
correctness (via `athens.posh/version`).

The assumption is that later this hybrid duality can be elided
and `athens.posh` refactored into a more robust data fetching layer.