feat(devtool): Add athens.tracing dev tool for profiling
Added athens.tracing which depends on tufte for profiling CLJS
functions. It's not a great integration, as currently Athens is on an
old version of re-frame-10x that does not have the appropriate
monkeypatchin hooks.

The assumption is that a developer will enable it by uncommenting the
`athens.tracing` in shadow-cljs preloads and checking the results in
console.log.