test(db): add Karma test for node search by title (#903)
I'd like to add a few Karma tests, so that I can later conveniently
change stuff closer to frontend without breaking it. The function
to search for node by title is a convenient first target (to just get
Karma tests set up and running) - it's simple, and I want to enhance
it a little bit soon.

(In particular, I want to make it a bit more fuzzy - e.g.,
if I search for "kiwi recipes", the page "Summer [[Kiwi]] and [[Melon]]
Recipes" should come up. Currently it does not, probably because it
does not ignore the `[[` / `]]`.)