enhance: Trigger loading of daily pages more eagerly
Fixes https://github.com/athensresearch/athens/issues/977.

We currently only load new daily pages when user scrolls to the very
bottom of the container also containing the "Earlier" header, and the
loading is also debounced by 500 ms, which adds a 500 ms delay before
the earlier pages actually start loading.

This commit shortens the debounce delay to just 100 ms, and triggers
the loading of earlier pages above the "Earlier" heading. Together this
should make loading of past pages feel a bit smoother, and it does seem
a bit smoother for me.