fix: don't flicker old state on save
This callback causes blocks that are saved to briefly flicker their old
content. I saw this happen with idle save, but I think I've also seen it
happen before briefly on up/down.

I don't think we need to do this because in
`athens.views.blocks.core/block-el` there is already init logic that
will reset string/previous when the new value comes in to the database.

Without using the callback in `transact-state-for-uid`, we can remove it
from the `:block/save` event since all other callsites had an empty callback.