feat: don't warn on exit while editing
This change was first introduced in
https://github.com/athensresearch/athens/pull/1899.

It made a lot of sense at the time because it was so easy to lose
content that was being edited.
But now we auto-save after 2s of inactivity in a block, so I don't think
it's as big of a problem anymore.

I do get the warning popup very often so I think now it's doing more
harm than good.