build: update shadow-cljs, cljs, tick
Tried running an Athens dev build on Windows and ran into https://github.com/thheller/shadow-cljs/issues/966, which is solved by updating Shadow-CLJS.

Shadow-CLJS and CLJS have to be updated together because that's just how Shadow-CLJS works.

Tick had to be updated because it broke when updating the other two.